WIIS-253: pagination fonctionne avec filtres

+++ src/Controller/FilterController.php
@@ -97,7 +97,6 @@ class FilterController extends AbstractController
                 ];
 
                 $result = [
-                    'reload' => $this->refArticleDataService->getRefArticleDataByParams($request->request),
                     'filterHtml' => $this->renderView('reference_article/oneFilter.html.twig', ['filter' => $filterArray])
                 ];
             } else {
@@ -118,14 +117,15 @@ class FilterController extends AbstractController
 
             if ($filterId) {
                 $filter = $this->filterRepository->find($filterId);
-                $em = $this->getDoctrine()->getManager();
-                $em->remove($filter);
-                $em->flush();
+
+                if ($filter) {
+                    $em = $this->getDoctrine()->getManager();
+                    $em->remove($filter);
+                    $em->flush();
+                }
             }
 
-            $data = $this->refArticleDataService->getRefArticleDataByParams();
-
-            return new JsonResponse($data);
+            return new JsonResponse();
         }
         throw new NotFoundHttpException("404");
     }
